requirements.txt
CTkListbox==1.4
customtkinter==5.2.2
darkdetect==0.8.0
grpcio==1.64.1
grpcio-tools==1.64.1
packaging==24.0
protobuf==5.27.1
setuptools==70.0.0
